+++ src/main/java/edu/harvard/iq/dataverse/api/DownloadInstanceWriter.java
@@ -642,7 +642,7 @@ public class DownloadInstanceWriter implements MessageBodyWriter<DownloadInstanc
         }
     }
     
-    // TODO: Return ".md" for "text/markdown" as well as other extensions in MimeTypeDetectionByFileExtension.properties
+    // TODO: Return file extensions for extensions in MimeTypeDetectionByFileExtension.properties and not supported by Tika
     private String getFileExtension(AuxiliaryFile auxFile) {
         String fileExtension = "";
         if (auxFile == null) {