+++ src/main/java/edu/harvard/iq/dataverse/api/AbstractApiBean.java
@@ -33,6 +33,7 @@ import edu.harvard.iq.dataverse.locality.StorageSiteServiceBean;
 import edu.harvard.iq.dataverse.metrics.MetricsServiceBean;
 import edu.harvard.iq.dataverse.search.savedsearch.SavedSearchServiceBean;
 import edu.harvard.iq.dataverse.settings.FeatureFlags;
+import edu.harvard.iq.dataverse.settings.JvmSettings;
 import edu.harvard.iq.dataverse.settings.SettingsServiceBean;
 import edu.harvard.iq.dataverse.util.BundleUtil;
 import edu.harvard.iq.dataverse.util.DateUtil;
@@ -324,7 +325,7 @@ public abstract class AbstractApiBean {
     }
 
     protected String getRequestApiKey() {
-        String headerParamApiKey = httpRequest.getHeader(DATAVERSE_KEY_HEADER_NAME);
+        String headerParamApiKey = httpRequest.getHeader(ApiConstants.DATAVERSE_KEY_HEADER_NAME);
         String queryParamApiKey = httpRequest.getParameter("key");
                 
         return headerParamApiKey!=null ? headerParamApiKey : queryParamApiKey;
@@ -463,11 +464,11 @@ public abstract class AbstractApiBean {
             }
         } else {
             String persistentId = id;
-            if (id.equals(PERSISTENT_ID_KEY)) {
-                persistentId = getRequestParameter(PERSISTENT_ID_KEY.substring(1));
+            if (id.equals(ApiConstants.PERSISTENT_ID_KEY)) {
+                persistentId = getRequestParameter(ApiConstants.PERSISTENT_ID_KEY.substring(1));
                 if (persistentId == null) {
                     throw new WrappedResponse(
-                            badRequest(BundleUtil.getStringFromBundle("find.dataset.error.dataset_id_is_null", Collections.singletonList(PERSISTENT_ID_KEY.substring(1)))));
+                            badRequest(BundleUtil.getStringFromBundle("find.dataset.error.dataset_id_is_null", Collections.singletonList(ApiConstants.PERSISTENT_ID_KEY.substring(1)))));
                 }
             }
             GlobalId globalId;
@@ -488,7 +489,7 @@ public abstract class AbstractApiBean {
                     fprLogService.logEntry(entry);
                 }
                 throw new WrappedResponse(
-                        notFound(BundleUtil.getStringFromBundle("find.dataset.error.dataset_id_is_null", Collections.singletonList(PERSISTENT_ID_KEY.substring(1)))));
+                        notFound(BundleUtil.getStringFromBundle("find.dataset.error.dataset_id_is_null", Collections.singletonList(ApiConstants.PERSISTENT_ID_KEY.substring(1)))));
             }
         }
         if (deep) {
@@ -586,11 +587,11 @@ public abstract class AbstractApiBean {
 
     protected DataFile findDataFileOrDie(String id) throws WrappedResponse {
         DataFile datafile;
-        if (id.equals(PERSISTENT_ID_KEY)) {
-            String persistentId = getRequestParameter(PERSISTENT_ID_KEY.substring(1));
+        if (id.equals(ApiConstants.PERSISTENT_ID_KEY)) {
+            String persistentId = getRequestParameter(ApiConstants.PERSISTENT_ID_KEY.substring(1));
             if (persistentId == null) {
                 throw new WrappedResponse(
-                        badRequest(BundleUtil.getStringFromBundle("find.dataset.error.dataset_id_is_null", Collections.singletonList(PERSISTENT_ID_KEY.substring(1)))));
+                        badRequest(BundleUtil.getStringFromBundle("find.dataset.error.dataset_id_is_null", Collections.singletonList(ApiConstants.PERSISTENT_ID_KEY.substring(1)))));
             }
             datafile = fileService.findByGlobalId(persistentId);
             if (datafile == null) {
@@ -633,8 +634,8 @@ public abstract class AbstractApiBean {
 
     protected DataverseRole findRoleOrDie(String id) throws WrappedResponse {
         DataverseRole role;
-        if (id.equals(ALIAS_KEY)) {
-            String alias = getRequestParameter(ALIAS_KEY.substring(1));
+        if (id.equals(ApiConstants.ALIAS_KEY)) {
+            String alias = getRequestParameter(ApiConstants.ALIAS_KEY.substring(1));
             try {
                 return em.createNamedQuery("DataverseRole.findDataverseRoleByAlias", DataverseRole.class)
                         .setParameter("alias", alias)
@@ -666,11 +667,11 @@ public abstract class AbstractApiBean {
         DatasetLinkingDataverse dsld;
         Dataverse linkingDataverse = findDataverseOrDie(linkingDataverseId);
 
-        if (datasetId.equals(PERSISTENT_ID_KEY)) {
-            String persistentId = getRequestParameter(PERSISTENT_ID_KEY.substring(1));
+        if (datasetId.equals(ApiConstants.PERSISTENT_ID_KEY)) {
+            String persistentId = getRequestParameter(ApiConstants.PERSISTENT_ID_KEY.substring(1));
             if (persistentId == null) {
                 throw new WrappedResponse(
-                        badRequest(BundleUtil.getStringFromBundle("find.dataset.error.dataset_id_is_null", Collections.singletonList(PERSISTENT_ID_KEY.substring(1)))));
+                        badRequest(BundleUtil.getStringFromBundle("find.dataset.error.dataset_id_is_null", Collections.singletonList(ApiConstants.PERSISTENT_ID_KEY.substring(1)))));
             }
             
             Dataset dataset = datasetSvc.findByGlobalId(persistentId);
@@ -1042,71 +1043,67 @@ public abstract class AbstractApiBean {
      *  HTTP Response methods *
     \* ====================== */
 
-    protected Response ok( JsonArrayBuilder bld ) {
-        return Response.ok(Json.createObjectBuilder()
-            .add("status", ApiConstants.STATUS_OK)
-            .add("data", bld).build())
-            .type(MediaType.APPLICATION_JSON).build();
-    }
-
-    protected Response ok( JsonArrayBuilder bld , long totalCount) {
-        return Response.ok(Json.createObjectBuilder()
-                        .add("status", ApiConstants.STATUS_OK)
-                        .add("totalCount", totalCount)
-                        .add("data", bld).build())
-                .type(MediaType.APPLICATION_JSON).build();
-    }
-
-    protected Response ok( JsonArray ja ) {
-        return Response.ok(Json.createObjectBuilder()
-            .add("status", ApiConstants.STATUS_OK)
-            .add("data", ja).build())
-            .type(MediaType.APPLICATION_JSON).build();
-    }
-
-    protected Response ok( JsonObjectBuilder bld ) {
-        return Response.ok( Json.createObjectBuilder()
-            .add("status", ApiConstants.STATUS_OK)
-            .add("data", bld).build() )
+    protected Response ok(JsonValue value, JsonValue message, Long totalCount) {
+        return Response.status(Response.Status.OK)
+            .entity(NullSafeJsonBuilder.jsonObjectBuilder()
+                .add(ApiConstants.STATUS_FIELD, ApiConstants.STATUS_OK)
+                .add(ApiConstants.MESSAGE_FIELD, message)
+                .add(ApiConstants.TOTAL_COUNT_FIELD, totalCount)
+                .add(ApiConstants.DATA_FIELD, value)
+                .build())
             .type(MediaType.APPLICATION_JSON)
             .build();
+    
     }
     
-    protected Response ok( JsonObject jo ) {
-        return Response.ok( Json.createObjectBuilder()
-                .add("status", ApiConstants.STATUS_OK)
-                .add("data", jo).build() )
-                .type(MediaType.APPLICATION_JSON)
-                .build();    
+    protected Response ok(JsonArrayBuilder bld) {
+        return ok(bld.build(), null, null);
     }
 
-    protected Response ok( String msg ) {
-        return Response.ok().entity(Json.createObjectBuilder()
-            .add("status", ApiConstants.STATUS_OK)
-            .add("data", Json.createObjectBuilder().add("message",msg)).build() )
-            .type(MediaType.APPLICATION_JSON)
-            .build();
+    protected Response ok(JsonArrayBuilder bld, long totalCount) {
+        return ok(bld.build(), null, totalCount);
+    }
+
+    protected Response ok(JsonArray ja) {
+        return ok(ja, null, null);
+    }
+
+    protected Response ok(JsonObjectBuilder bld) {
+        return ok(bld.build(), null, null);
     }
     
-    protected Response ok( String msg, JsonObjectBuilder bld  ) {
-        return Response.ok().entity(Json.createObjectBuilder()
-            .add("status", ApiConstants.STATUS_OK)
-            .add("message", Json.createObjectBuilder().add("message",msg))     
-            .add("data", bld).build())      
-            .type(MediaType.APPLICATION_JSON)
-            .build();
+    protected Response ok(JsonObject jo) {
+        return ok(jo, null, null);
+    }
+
+    protected Response ok(String msg) {
+        // An instance may opt out of using the old {data:{message:"$msg"}} way.
+        // This is a highly used response builder, which is why this is an experimental opt-in change!
+        // TODO: This will be removed in a future version.
+        if (FeatureFlags.UNIFY_API_RESPONSE_MESSAGE_STYLE.enabled()) {
+            return ok(null, Json.createValue(msg), null);
+        } else {
+            return ok(Json.createObjectBuilder().add("message", msg).build(), null, null);
+        }
+    }
+    
+    protected Response ok(String msg, JsonObjectBuilder bld) {
+        // Legacy mode returns message as nested object for backward compatibility with integrations that worked around the bug.
+        // This is a scarcely used way to build a response, mostly relevant to admins, which is why we make it opt-out.
+        // TODO: This will be removed in a future version.
+        if (JvmSettings.LEGACY_API_RESPONSE_MESSAGE_STYLE.lookupOptional(Boolean.class).orElse(false)) {
+            return ok(bld.build(), Json.createObjectBuilder().add(ApiConstants.MESSAGE_FIELD, msg).build(), null);
+        } else {
+            return ok(bld.build(), Json.createValue(msg), null);
+        }
     }
 
     protected Response ok( boolean value ) {
-        return Response.ok().entity(Json.createObjectBuilder()
-            .add("status", ApiConstants.STATUS_OK)
-            .add("data", value).build() ).build();
+        return ok(value ? JsonValue.TRUE : JsonValue.FALSE, null, null);
     }
 
     protected Response ok(long value) {
-        return Response.ok().entity(Json.createObjectBuilder()
-                .add("status", ApiConstants.STATUS_OK)
-                .add("data", value).build()).build();
+        return ok(Json.createValue(value), null, null);
     }
 
     /**
@@ -1131,8 +1128,8 @@ public abstract class AbstractApiBean {
     protected Response created( String uri, JsonObjectBuilder bld ) {
         return Response.created( URI.create(uri) )
                 .entity( Json.createObjectBuilder()
-                .add("status", "OK")
-                .add("data", bld).build())
+                .add(ApiConstants.STATUS_FIELD, ApiConstants.STATUS_OK)
+                .add(ApiConstants.DATA_FIELD, bld).build())
                 .type(MediaType.APPLICATION_JSON)
                 .build();
     }
@@ -1140,15 +1137,15 @@ public abstract class AbstractApiBean {
     protected Response accepted(JsonObjectBuilder bld) {
         return Response.accepted()
                 .entity(Json.createObjectBuilder()
-                        .add("status", STATUS_WF_IN_PROGRESS)
-                        .add("data",bld).build()
+                        .add(ApiConstants.STATUS_FIELD, ApiConstants.STATUS_WF_IN_PROGRESS)
+                        .add(ApiConstants.DATA_FIELD, bld).build()
                 ).build();
     }
     
     protected Response accepted() {
         return Response.accepted()
                 .entity(Json.createObjectBuilder()
-                        .add("status", STATUS_WF_IN_PROGRESS).build()
+                        .add(ApiConstants.STATUS_FIELD, ApiConstants.STATUS_WF_IN_PROGRESS).build()
                 ).build();
     }
 
@@ -1163,8 +1160,8 @@ public abstract class AbstractApiBean {
     protected Response badRequest(String msg, Map<String, String> fieldErrors) {
         return Response.status(Status.BAD_REQUEST)
                 .entity(NullSafeJsonBuilder.jsonObjectBuilder()
-                        .add("status", ApiConstants.STATUS_ERROR)
-                        .add("message", msg)
+                        .add(ApiConstants.STATUS_FIELD, ApiConstants.STATUS_ERROR)
+                        .add(ApiConstants.MESSAGE_FIELD, msg)
                         .add("fieldErrors", Json.createObjectBuilder(fieldErrors).build())
                         .build()
                 )
@@ -1197,7 +1194,7 @@ public abstract class AbstractApiBean {
     }
 
     protected Response authenticatedUserRequired() {
-        return error(Status.UNAUTHORIZED, RESPONSE_MESSAGE_AUTHENTICATED_USER_REQUIRED);
+        return error(Status.UNAUTHORIZED, ApiConstants.RESPONSE_MESSAGE_AUTHENTICATED_USER_REQUIRED);
     }
 
     protected Response permissionError( PermissionException pe ) {
@@ -1222,9 +1219,9 @@ public abstract class AbstractApiBean {
 
     protected static Response error( Status sts, String msg ) {
         return Response.status(sts)
-                .entity( NullSafeJsonBuilder.jsonObjectBuilder()
-                        .add("status", ApiConstants.STATUS_ERROR)
-                        .add( "message", msg ).build()
+                .entity(NullSafeJsonBuilder.jsonObjectBuilder()
+                        .add(ApiConstants.STATUS_FIELD, ApiConstants.STATUS_ERROR)
+                        .add(ApiConstants.MESSAGE_FIELD, msg ).build()
                 ).type(MediaType.APPLICATION_JSON_TYPE).build();
     }
 }