+++ doc/sphinx-guides/source/api/changelog.rst
@@ -25,6 +25,8 @@ v6.12
 
   - **/api/admin/index/perms/{id}**
 
+- Whether file extensions are included in the "Content-disposition" header returned when downloading auxiliary files depends on whether the relevant format is one of the Mimetypes supported in Tika. A recent update to the version of Tika has added new mimetypes, including "text/markdown", and auxiliary files with these types now have a a file extension included (e.g. ".md" in this case) in the header.
+
 v6.11
 -----